Ignore errors on parallel runs
Currently if you set concurrency to 1 with 2 test files, if 1 test file fails it exists the process. This may be fine but it makes a lot more sense to make this an option. For example I want to run only 5 processes and allow the tests to fail (they are recording their own xunit reports) and continue on to the next file.  Adding a configuration option ignoreErrors.